+++ game-core/game-core-entity/src/main/java/com/fulu/game/core/entity/DynamicFile.java
@@ -18,15 +18,15 @@ public class DynamicFile implements Serializable {
 	private static final long serialVersionUID = 1L;
 	
 	//
-	private Long id;
+	private Integer id;
 	//动态id
-	private Long dynamicId;
+	private Integer dynamicId;
 	//文件链接地址
 	private String url;
 	//文件类型(1：图片；2：视频，由于同一条动态不能同时选图片和视频，因此将此属性放到主表中)
 	private Integer type;
 	//播放次数（视频才会有）
-	private Long playCount;
+	private Integer playCount;
 	//创建时间（预留）
 	@JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss", timezone = "GMT+8")
 	private Date createTime;
